requirements.txt
pyTelegramBotAPI==4.0.1
RPi.GPIO==0.7.0
